--- src/main/java/lotto/LottoNumberGenerator.java
@@ -1,7 +1,0 @@
-package lotto;
-
-import java.util.Set;
-
-interface LottoNumberGenerator {
-    Set<Integer> generateNumbers();
-}
