+++ src/test/java/lotto/LottoTest.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.domain.Lotto;
+import lotto.domain.LottoRank;
+import lotto.domain.LottoStatistics;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
