+++ src/test/java/lotto/ProfitCalculatorTest.java
@@ -1,5 +1,6 @@
 package lotto;
 
+import lotto.domain.ProfitCalculator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
